Adjusted secure Nginx parameters
The goal will be to maintain A+, 100% ratings across the board on Qualys SSL Labs SSL Test (and potentially other tests). Made changes to the ciphers and curves used, also added "server_tokens off;" and "add_header Referrer-Policy "no-referrer-when-downgrade" always;". Will have to wait on ability to test LetsEncrypt to test these changes and verify the results.