Update guac-install.sh
Fixed Nginx proxy_pass parameter to http as traffic between Nginx and Tomcat are in the plain. This is with Nginx acting as a reverse proxy. Caused 502 errors in testing with LetsEncrypt cert but not with self signed. I am guessing this is because the slef signed worked for Tomcat and Nginx whereas the Letsencrypt cert worked only for Nginx and so encrypted traffic from Nginx to Tomcat fails.