Update guac-install.sh
Removed the variables GUACSERVER_HOSTNAME and GUACSERVER_HOSTNAME_DEF. Changed the nginx confs to use "localhost" statically for the proxy_pass lines. Updated the URI prompt to reflect needing to start and end with a slash.

For the DOMAIN_NAME prompts, used for the server_name in the nginx confs, updated the wording to reflect hostname or public domain and set the default to localhost. For now have commented out the prompt for the domain name in the Lets Encrypt menu. May delete that prompt later if its obviously not required.