Update guac-install.sh
Implemented a new menu thats much more dynamic and organized into sub groups. Includes the ability to review and alter settings or start the script from the beginning prior to running the script. Also includes options to cancel and exit the script.

In recreating the majority of the menu, I aim to account for adding the ability to install other extensions in addition to the LDAP extension. I will also provide 3 options for Nginx configuration according to "security level".

This is far from complete, with many functions/segments of code needing attention to accommodate the new menus.

Testing will need to be done to ensure the new menu system has not caused any regressions in previously working functionality.