Update guac-install.sh
Reverted default for MYSQL_CON_VER to 5.1.47 from 8.0.15 as 8x was causing issues with mariaDB that resulted in blank login screen (see issue #38. Feel free to try 8.x but verify it works before relying on it. 8.x seems to work for me with a Lets Encrypt SSL cert and public domain, but not in testing with a self signed cert and local access via IP/hostname.