Update guac-install.sh
Added options menus and summary menus for the GUAC_SERVER_IP variable to allow manually setting it or using the default pulled by the hostname -I command. This address is used internally on the server for the proxy_pass parameters in the Nginx conf files, in the Tomcat configuration and in some echo'd output. This should allow more control when using a multiple-NIC server with multiple IP address's in case the one returned by hostname -I is not the desired one.

This should not be confused with the server_name in the Nginx confs or as the name or IP used to access the servers web UI/login page or for other purposes.