Update guac-install.sh
Removed deprecated variable for the parameters assigned to guacd-hostname and mysql-hostname in the guacamole.properties file. These are statically set to localhost now. As per Guac documentation, when left blank they default to local host and since the variable was never assigned they were blank. As the script configures both guacd and mariadb locally, there is not reason for them not to be localhost.