Update guac-install.sh
Re-wrote the entire stdout/echo'd message during script run time. This should be more informative while also being less cluttered/busy. I have updated these messages to more often and more accurately the progress spinner when it makes sense to show it. The progress spinner has been updated to hopefully display the word "-done-" when complete instead of a slash. Moved some blocks of code to other functions to better group progress messages. Removed the pv function from installed packages.

Will need to test to ensure out put is as desired and that it has not adversely affected the actions of the script.