Update guac-install.sh
Added a trap for ERR SIGINT and SIGQUIT along with an error handler function. This covers the installation code but does not error handle on the menus or summaries. I have removed all "|| error 1" lines from individual commands in favor of using the trap. The result should be that if an error occurs during installation (or manual stop like ctrl+c) that the script will output to stdout the error, line number in the script, command and exit code then cease running. Also changed the script to finish with exit 0 instead of exit 1 (why I did this prior, the world may never know)