Update guac-install.sh
Added chk_installed function to replace rpm -qa | grep lines for checking if packages are installed like epel-release and rpmfusion. This function checks if the package is installed using rpm -q package_name and then returns the retval without triggering the trap. This way the script does not exit when the package is not already installed.