Update guac-install.sh
Update to display variable values in error message instead of the variable name. This should make it easier to tell if a parameter was entered incorrectly causing it to fail. I am testing this to make sure it works and to see if its a safe method in the context of this script to utilize.