Update guac-install.sh
Added prompt to enable CSP to the Nginx menu. Added CSP selection to Nginx summary. Added CSP line to guacamole_ssl.conf with logic to include it commented out if the selection was to disable it. Updated OCSP Stapling logic in guacamole_ssl.conf to included a commented line for OCSP parameters if its chosen to not use it, instead of previously not including the lines in the configuration at all. This should make it easier for users to enable aspects of the Nginx SSL config later by uncommenting lines.