Update guac-install.sh
Reverted changes to install function order and instead removed wget from the main yum install line and moved it to the block for libjpeg-turbo to ensure wget is install ahead of anything dependent on it. Re-arranging the functions caused other issues and would have required more additional lines of code for the same result.