Update guac-install.sh
Added new variables to store minor version and release portion of OS version. Using minor version to handle freerdp install on 7.6 and lower vs 7.7 and higher (not accounting for RHEL/CentOS 8.x yet). If the OS is 7.7 or greater, freerdp 1.x is installed from CentOS-Vault, an exclude line is added to yum.conf for freerdp to prevent future updates and then all other packages are installed.