Update guac-install.sh
Updated the CentOS-Vault repo portion to get gpg key from the site vs from the expected file on a CnetOS machine (applies to RHEL host only). Be aware that as it is in CentOS, the baseurl is HTTP, not HTTPS. While HTTPS works when going to the site in a browser, it appears to break the repo when trying to install packages. Likewise, the gpg key is also using an HTTP url, HTTPS also breaks the ability to acquire it from yum. Evaluate the risk for yourself before using this as is.