Update guac-install.sh
Cleaned up the code related to freerdp for both RHEL and CentOS. It appears the freerdp 1.0.2 packages are available from the 1st party RHEL repos as well. Added the package yum-versionlock which is a yum plugin to essentially exclude packages from updates. In this case by adding all non-version freerdp 1.0.2 to the list of packages to exclude.