Update guac-install.sh
versionlock on freerdp was not working as expected. The initial install would work but running yum update would still update freerdp to 2.x on CentOS. This was because the repo with the freerdp 1.x packages was being enabled for RHEL earlier in the code, but the CentOS specific repo with the older freerdp version was only temporarily enabled at the time of running the script.

This fix will keep the repo for 7.6 with freerdp 1.x enabled persistently so that versionlock is effective.