Update guac-install.sh
Added code to create a user and group for guacd with nologin but with a home dir as required by freerdp 2.x. The guacd service will be run from this new user, "guacd", whose name and group can be changed by the GUACD_USER variable prior to running the script. This user is created as a service account using the -r switch of useradd.

RDP connections should work now, depending on client the connection may need certain options selected to work even if it did not in a prior version of Guac.