Update guac-install.sh
Added 2 variables, NAME_SERVERS and NAME_SERVERS_DEF to store DNS name servers for the resolver parameter of the Nginx config used with OCSP. This is to fix a problem with Firefox (and possibly other browsers in the future) not loading Guacamole when the parameter is missing and is using OCSP.

The variable is populated with the name servers of the host machine (the Guacamole server the script is run on) from the /etc/resolv.conf file. All nameserver lines from the file are pulled and formatted in a space delimited single line string. The default if the file is empty is to use the NAME_SERVERS_DEF value which can be changed in the same area as other common variables towards the top of the script. I have set it to the 4 IP's for Cloudflares 1.1.1.1 DNS service. 2 IP's for IPv4 and 2 IP's for IPv6.