Update guac-install.sh
Seems that the repo file for Nginx has an additional line in its current iteration from the Nginx.org site. Added this line to the .repo file in the script. Also updated MySQL Connector J from 8.0.19 to 8.0.20. Removed dialog package as I think its a remnant of an old approach at doing something in the script no longer done.